moved WideMapScreen styles to constants file
