+++ constants/Map/MapSearchBarComponent.js
