+++ constants/Map/WideMapScreen.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    paddingBottom: '5%',
+    justifyContent: 'flex-start',
+    alignItems: 'center'
+  },
+  mapSearchBar: {
+    zIndex: 1000
+  }
+});
+
+export default styles;
