+++ screens/maps/WideMapScreen.js
@@ -1,8 +1,10 @@
-import React from "react";
-import { View, StyleSheet } from "react-native";
-import WideMap from "./WideMap";
-import MapSearchBarComponent from "../../components/Search/MapSearchComponent";
-import BackButton from "../../components/BackButton";
+import React from 'react';
+import { View } from 'react-native';
+
+import WideMap from './WideMap';
+import MapSearchBarComponent from '../../components/Search/MapSearchComponent';
+import BackButton from '../../components/BackButton';
+import styles from '../../constants/Map/WideMapScreen';
 
 class WideMapScreen extends React.Component {
   constructor(props) {
@@ -10,11 +12,11 @@ class WideMapScreen extends React.Component {
   }
   static navigationOptions = ({ navigation }) => {
     return {
-      title: "Organization Locations",
+      title: 'Organization Locations',
       headerStyle: {
-        backgroundColor: "#323338"
+        backgroundColor: '#323338'
       },
-      headerTintColor: "#fff",
+      headerTintColor: '#fff',
       headerLeft: () => <BackButton navigation={navigation} />
     };
   };
@@ -22,25 +24,13 @@ class WideMapScreen extends React.Component {
   render() {
     return (
       <View style={styles.container}>
-        {/* <MapScreenHeader /> */}
-        <MapSearchBarComponent style={styles.mapSearchBar} />
+        <MapSearchBarComponent
+        // style={styles.mapSearchBar}
+        />
         <WideMap navigation={this.props.navigation} />
       </View>
     );
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    flexDirection: "column",
-    paddingBottom: "5%",
-    justifyContent: "flex-start",
-    alignItems: "center"
-  },
-  mapSearchBar: {
-    zIndex: 1000
-  }
-});
 
 export default WideMapScreen;
